Fixes following elements of design review:
A1) Missing the “Click on the block to see sub-functions for (function name)” text.
C1) 30px of margin/padding between the header and text block.
D1) Block titles/totals should have a drop shadow until the block they belong to is hovered over and switch to the dark text.
B2) This text shouldn’t be visible when viewing the subfunctions.
C2) 30px of margin/padding between these sections.
D2) 15px of margin/padding between the bar and the up arrow.
F2) The source text should be left aligned with the treeMap above.
A3) The opacity of the navigation block should be 100% on hover.