[DEV-6114] fixing fragile test
Test was failing with new node for some reason?
Test required the array to have specific order. Instead of relying on deep equality
comparison of arrays being equal (which requires the sort order to be exact), we are
using .some() to just check and make sure the array has the correct
items in it.