[DEV-6377] fiscal year & period + test updates:
(1) TimeFilter.jsx
a. capturing period/fy in url
b. validating period/fy in url
c. customHook for fetching latest submission
data
d. styling hack on order of period dropdown

NOTE on (d):
This hack is so that we can use the ~ sibling-element
selector that only selects "younger" siblings. We
order the items in the dropdown so this is possible, then re-order them
via css.

(2) test config updates:
- new component to register icons in app and in test
- adding to modulePaths array