[DEV-6662] Glossary Trigger Edge Case Logic:
When we're on a page that has query params, triggering
the glossary is semi more complex. The glossary was previously
the only place where we were using query  params. When we trigger
the glossary we are updating the url like `/?glossary=glossary-term`,
updating redux to set that as the active term, and then removing the
query params from the url.

This commit introduces a helper called getNewUrlForGlossary which
is called when seeking to trigger the glossary on a page with query
params. It provides a new url with the glossary param
appended to any existing params.

Tests included.