[DEV-6449] refactoring logic for fetching new data & custom hook "usePagination".
Refactoring logic for fetching new data:
-- (a) trigger a request when page changes
-- (b) go to page one when sort/limit/search changes
-- (c) when sort/limit/search changes and on page one,
refetch, but **only** if submissions/publications.length to
avoid fetch on render. This was necessary to get the tests
to pass.

Refactoring "usePagination":
-- exposing methods that update specific properties in
the paginationObj { page, limit, totalItems }, rather than
one method that updates every property with a new object.
-- only updating totalItems in API response then() block.
Updating the page/limit is redundant.