[DEV-6748] Refactor adv search to use queryParams:
-- SearchContainer: (A) use useLocation.search & useParams to
expect a hash as a query param (/search/?hash=xyz) or in the
path (/search/:hash); (B) on "reset search" if using hash,
manually disable the download button (when not using hash, a
re-render/mount automatically sets)

-- searchHelper: refactor isSearchHashReady
to expect an object w/ either pathname or search
containing a hash.

-- ResultsTableContainer: consumer of isSearchHashReady:

didMount handles case where it's in the pathname,
didUpdate handles case where its in the search keyspace.
Was able to remove obsolete condition as well where we're
looking at the filters directly to detect a change.

TESTS:
-- searchHelper: new tests for refactor of isSearchHashReady
-- SearchContainer: WIP; refactored to mock for queryParams
and use history.push instead of history.replace