[DEV-6748] refactor areFiltersEqual to use both lodash.isEqual and immutable.is().
areFiltersEqual depended on the Immutable.is() operator. It is
constrained to primitive (non object/array) values or
immutable data structures for accurate nested object
comaprisons. We were using an immutable.Record w/ non
immutable, non-primitive values for the checkboxTree filters.
This resulted in wrong returns on the immutable.is()
invocations inside areFiltersEqual.

This refactor adds a new path for non-immutable objects in
our redux.appliedFilters and redux.filters keyspace
that allows for an accurate comparison using
lodash.isEqual. Tests were added to throughouly demonstrate this.

We add a new invocation of this fn, areFiltersEqual, inside
SearchContainer to handle the case where a user is on adv
search and clicks the pre-poulate adv search link in the COVID Modal.