[DEV-6721] Dynamic DEFCodes for COVID DS&M pt I:
WithDefCodes implements an existing pattern set by WithLatestFy.

It exports a reusable hook that updates the DEFC redux keyspace when not
defined via the api call, and a HOC that does the same and passes the
redux values to the children.

Remaining commits will be implementing this HOC/hook everywhere def
codes are used and finally on the DS&M pg where we currently are relying
on hard-coded def_code values.