[tests-agency-details] About the Data, Agency Details Tests:
This commit seeks to offer validation that we are not making
duplicate API requests. It appears in the browser we are not making
duplicate requests already, but our tests do not agree with that
conclusion.

Container refactor:
- Introducing prevPage to use in conditionals within useEffect to offer
more robust logic to confirm when and how to make a request

Tests refactor:
- Changing the expect statement to reflect what the user would
expect: first render, 1 request; change sort field, 2 requests; change
sort order, 2 requests