[DEV-6871] addChildrenAndPossiblyPlaceholder:
handling edge case where areChildrenPartial is false
but there are grand children that are not genuine.

Introducing the variable hasPlaceholderGrandChildren
to acknowledge this case and handle it by ensuring
those grandChildren are hidden and we dont have a placeholder
show up in the search results w/o the "hide" class.

Test case added to checkboxTreeHelper-test.