[DEV-6739] invoking fetchTotals again ...
611cb3f35d65a5e1ff0f7ad2ca6b325660582fe9

Above commit refactored useEffect logic to reset to page one
when the following things changed:

 - fiscal year
 - period
  - sort direction
   - sort order

I unintentionally broke the logic where we fetch the totals and populate the
Percent of Total Federal Budget column. This was caused by removing
selectedFy and selectedPeriod from the dependency array of the useEffect
where we invoke fetchTotals.

On the first render, selectedFy and selectedPeriod are undefined, so we
have to listen for when they become defined, then we call fetchTotals.

This commit moves the call to fetchTotals back to the
useEffect that fires on fiscal year of fiscal period change.

Test added to cover scenario of first render where (a) fy and period are
undefined, (b) they get defined, (c) we call fetchTotals, and (d)
we call fetchTableData.