[DEV-6805] About the data row and modal date discrepancy: Showing UTC MM-dd-yyyy format in modal and EST MM-dd-yy format in table.
This date string 2020-11-14T01:49:36.677591Z was surfacing a bug where
in the table it was formatted as 11-13-2020 and in the modal, 11-14-2020

Both dates are right. The question becomes what do we show to the user,
given a UTC timestamp from the API? The answer is we show this date to
the user with the relevant time-zone based offset applied. We do this via
the browser knowing the user's timeszone an npm package called dates-fn
which applies the offset for us.

In Charelston, SC, we are in EST. Which is a minus 5 hour offset from
UTC. So the timestamp becomes 2020-11-13T20:49:36.677591-05:00
or 2020-11-13 at 08:49PM EST when the offset is applied.

Deleting code which was a custom implementation of the date-fns format
method.

+Test config that sets default timezone to America/New_York
+Test for renderDeadlines