[DEV-4521] new hook: useValideTimeBasedQueryParams
Q: What does it do?
A: Takes query params and returns them as state after validating them in
comparison with the latest available data. It also updates the URL for
you.

Improvements:
1. If only one param is invalid, it will only update the invalid param.
In the case where you provide a valid FY and invalid period, however,
the period is set to the latest absolute period (Based on actual current
FY) rather than to the latest period available within whatever valid FY
was provided. I'm not gonna worry about this ATM.

2. if a period param is invalid because its too small, like a zero or
negative we are now handling that with an added condtion (> 0)
in the helper isPeriodVisible

Tests added for custom hook to cases where:

- when params are undefined
- params are defined but invalid