[DEV-6706] Account Download | DEFC User Selection validation logic update:
if a user only has selected accountBalances for file type, DEFC filter
selections are not relevant. In the first delivery of this ticket, we made
it so when this was the only selection the DEFC checkbox tree was
disabled.

This delivery ensures the user selection section removes DEFC from
the selected options when relevant. Specifically, this handles the edge
case where a user selects DEFC, then removes the file type down to
only accountBalances. In this case, we not only disable the tree but
we also remove the DEFC from the selected options.