[DEV-6691] Glossary Query Param Bug:
-- Rename params in helper to make it a bit more readable. Also adds
an additional condition; namely, if there are existing query params
but the existing url does not contain an appended '/'. This amounts
to a bug in the helper itself because both 'existing/url?param=x' and
'existing/url/?param=x' are technically valid and we need to handle
both.

- adds some test cases, and cleans up the existing ones to be a bit
more readable as well.

- GlossaryLink now calls new helper

- GlossaryListener now updates the URL with retained query params
from the previous URL