Updated advanced search Filters tooltip
The tooltip for filters should now change based on the screen size.

 At medium screen sizes (992px) and larger the tooltip should display to the right of the icon.

For the tooltip on advanced search filter sections less than medium screen(992px) to tablet screen (768px) the tooltip should be on the bottom of the icon with an arrow.

The advanced search page filters tooltip for anything less than tablet screen (768) the tooltips should be below the tooltip and slightly to the left.